formateventdatetime util
