+++ lib/utils.ts
@@ -44,3 +44,27 @@ export const formatPhoneNumber = (
   // Otherwise, return the original
   return phoneNumber
 }
+
+export const formatEventDateTime = (datetime: string | null) => {
+  if (!datetime) return 'Date TBD'
+
+  try {
+    const date = new Date(datetime)
+    const dateStr = date.toLocaleDateString('en-US', {
+      weekday: 'long',
+      year: 'numeric',
+      month: 'long',
+      day: 'numeric',
+      timeZone: 'America/Chicago',
+    })
+    const timeStr = date.toLocaleTimeString('en-US', {
+      hour: 'numeric',
+      minute: '2-digit',
+      hour12: true,
+      timeZone: 'America/Chicago',
+    })
+    return { dateStr, timeStr: `${timeStr} CT` }
+  } catch {
+    return 'Invalid Date'
+  }
+}
